Revert "cpufreq: lower suspend freq to 600MHz"
It's more for documentation, because this value is recalculated at
startup anyway.

RK3188 needs some minimal voltage to wake up and this voltage is higher
than required for it to operate. Default is 1V, but my tablet managed to
wake up even at 0.875V, but nothing less.

So it's not about suspend freq, but more about suspend voltage that
needs to be set before going into suspend

This reverts commit 3e1c332beed30cd51f5fffd7961627d2876168b0.